NP-8443 Move admin components
