+++ src/pages/basic_data/app_admin/CentralImportPage.tsx
@@ -1,7 +1,7 @@
 import { Typography } from '@mui/material';
 import { useTranslation } from 'react-i18next';
 
-export const CentralImport = () => {
+export const CentralImportPage = () => {
   const { t } = useTranslation('basicData');
   return <Typography variant="h2">{t('central_import')}</Typography>;
 };
