+++ src/pages/basic_data/institution_admin/AddAffiliationPanel.tsx
@@ -4,18 +4,18 @@ import LooksTwoIcon from '@mui/icons-material/LooksTwo';
 import { ErrorMessage, Field, FieldProps, useFormikContext } from 'formik';
 import { DatePicker } from '@mui/lab';
 import { useSelector } from 'react-redux';
-import { StyledCenterContainer } from '../../components/styled/Wrappers';
-import { CristinApiPath } from '../../api/apiPaths';
-import { LanguageString } from '../../types/common.types';
-import { useFetch } from '../../utils/hooks/useFetch';
-import { getLanguageString } from '../../utils/translation-helpers';
-import { datePickerTranslationProps } from '../../themes/mainTheme';
-import { getNewDateValue } from '../../utils/registration-helpers';
-import { RootStore } from '../../redux/reducers/rootReducer';
-import { Organization } from '../../types/organization.types';
-import { useFetchResource } from '../../utils/hooks/useFetchResource';
-import { getSortedSubUnits } from '../../utils/institutions-helpers';
-import { AddEmployeeData } from './AddEmployee';
+import { StyledCenterContainer } from '../../../components/styled/Wrappers';
+import { CristinApiPath } from '../../../api/apiPaths';
+import { LanguageString } from '../../../types/common.types';
+import { useFetch } from '../../../utils/hooks/useFetch';
+import { getLanguageString } from '../../../utils/translation-helpers';
+import { datePickerTranslationProps } from '../../../themes/mainTheme';
+import { getNewDateValue } from '../../../utils/registration-helpers';
+import { RootStore } from '../../../redux/reducers/rootReducer';
+import { Organization } from '../../../types/organization.types';
+import { useFetchResource } from '../../../utils/hooks/useFetchResource';
+import { getSortedSubUnits } from '../../../utils/institutions-helpers';
+import { AddEmployeeData } from './AddEmployeePage';
 
 interface Position {
   id: string;
