+++ src/pages/basic_data/institution_admin/FindPersonPanel.tsx
@@ -5,15 +5,15 @@ import { useTranslation } from 'react-i18next';
 import { ErrorMessage, Field, FieldProps, useFormikContext } from 'formik';
 import { useState, useCallback, useEffect } from 'react';
 import { Box } from '@mui/system';
-import { convertToFlatCristinUser } from '../../utils/user-helpers';
-import { CristinApiPath } from '../../api/apiPaths';
-import { authenticatedApiRequest } from '../../api/apiRequest';
-import { CristinUser } from '../../types/user.types';
-import { isSuccessStatus } from '../../utils/constants';
-import { StyledCenterContainer } from '../../components/styled/Wrappers';
-import { AddEmployeeData, emptyUser } from './AddEmployee';
-import { AffiliationHierarchy } from '../../components/institution/AffiliationHierarchy';
-import { getLanguageString } from '../../utils/translation-helpers';
+import { convertToFlatCristinUser } from '../../../utils/user-helpers';
+import { CristinApiPath } from '../../../api/apiPaths';
+import { authenticatedApiRequest } from '../../../api/apiRequest';
+import { CristinUser } from '../../../types/user.types';
+import { isSuccessStatus } from '../../../utils/constants';
+import { StyledCenterContainer } from '../../../components/styled/Wrappers';
+import { AddEmployeeData, emptyUser } from './AddEmployeePage';
+import { AffiliationHierarchy } from '../../../components/institution/AffiliationHierarchy';
+import { getLanguageString } from '../../../utils/translation-helpers';
 
 export const FindPersonPanel = () => {
   const { t } = useTranslation('basicData');
